Add contract call decoder
This commit adds a class that takes in the raw bytes of a contract call
and returns a readable explanation of the function call that it
represents. This is helpful for many situations where one needs to
determine the intent of a transaction when they only have the raw
transaction bytes.